backend/requirements.txt
# Core web framework
fastapi>=0.110.0,<1.0.0

# ASGI server for local development and deployment
uvicorn[standard]>=0.29.0,<0.30.0

psycopg2-binary>=2.9.9,<3.0.0

# Data validation and settings management
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.2.1,<3.0.0

# Database ORM
SQLAlchemy>=2.0.29,<3.0.0

# JWT handling
python-jose[cryptography]>=3.3.0,<4.0.0

# Testing dependencies
pytest>=8.1.0,<9.0.0
httpx>=0.27.0,<0.28.0
